Moved RealmBaseAdapter to io.realm.android
