+++ realm/src/androidTest/java/io/realm/entities/RealmAdapter.java
@@ -22,7 +22,7 @@ import android.view.ViewGroup;
 import android.widget.ListAdapter;
 import android.widget.TextView;
 
-import io.realm.RealmBaseAdapter;
+import io.realm.android.RealmBaseAdapter;
 import io.realm.RealmResults;
 
 public class RealmAdapter extends RealmBaseAdapter<AllTypes> implements ListAdapter {
